Step 2 (drift fix): delete duplicate model.CovariateProfile; consolidate on api.covariate.CovariateProfile
Resolves the CovariateProfile duplicate flagged in
DIR-PACKAGE-DRIFT-FIX-punit step 2. The canonical home is
api.covariate.CovariateProfile (Map<String, String> of name → label,
LinkedHashMap-backed for insertion-order preservation). The deleted
model.CovariateProfile carried richer typed-value (CovariateValue)
machinery and hash-computation methods; investigation showed the
typed wrapper was a sealed interface with a single StringValue
implementation, the hash methods were exercised only by their own
unit tests, and the production engine consumers used the model
shape only as a String container. Folding the engine path onto the
api shape is therefore non-lossy.

Touched:
- spec.registry.SpecificationLoader: replace CovariateProfile.Builder
  accumulator with a LinkedHashMap; coerce parseCovariateValue's
  CovariateValue to String via toCanonicalString() at the put site;
  finalize via api CovariateProfile.of(map).
- spec.model.ExecutionSpecification: replace the implicit-covariate
  merge builder loop with a LinkedHashMap that absorbs warmup /
  maxConcurrent then the carried profile's values(), finalised via
  CovariateProfile.of(merged).
- model.BaselineProvenance: switch the import; replace the
  orderedKeys()/get()/toCanonicalString() loop with a values()
  Map.Entry iteration.
- model.BaselineProvenanceTest: switch the import; replace
  CovariateProfile.builder() with LinkedHashMap + CovariateProfile.of().
- Delete model/CovariateProfile.java and its 17-test test class
  (the dead-in-production hash methods went with them).
- docs/DOMAIN-ONTOLOGY.md: mark the duplicate resolved in the
  Covariate-as-identity invariant note and the "Drift to fix" item.

Architecture-test bookkeeping (PackageStructureArchitectureTest):
- noTopLevelModelPackage rule's freeze-store shrunk by 2 entries
  (CovariateProfile main + test class). Other entries unchanged.
- engineUnderInternal rule's freeze-store unchanged (no movement
  into engine/).

CovariateValue itself is unchanged in this commit; its move to
api/covariate/ is owned by step 3a.

Closes ArchUnit rule contribution: noTopLevelModelPackage (one
class removed; rule still has further entries pending step 3).
