+++ punit-core/archunit_store/90083594-5684-45be-a3cf-87dc0d2f9509
@@ -1,7 +1,5 @@
 Class <org.javai.punit.model.BaselineProvenance> does reside in a package 'org.javai.punit.model..' in (BaselineProvenance.java:0)
 Class <org.javai.punit.model.CovariateDeclaration> does reside in a package 'org.javai.punit.model..' in (CovariateDeclaration.java:0)
-Class <org.javai.punit.model.CovariateProfile$Builder> does reside in a package 'org.javai.punit.model..' in (CovariateProfile.java:0)
-Class <org.javai.punit.model.CovariateProfile> does reside in a package 'org.javai.punit.model..' in (CovariateProfile.java:0)
 Class <org.javai.punit.model.CovariateValue$StringValue> does reside in a package 'org.javai.punit.model..' in (CovariateValue.java:0)
 Class <org.javai.punit.model.CovariateValue> does reside in a package 'org.javai.punit.model..' in (CovariateValue.java:0)
 Class <org.javai.punit.model.DayGroupDefinition> does reside in a package 'org.javai.punit.model..' in (DayGroupDefinition.java:0)
